# Push filter function from Python to SQL