--- a/tests/utils/test_search_utils.py
+++ b/tests/utils/test_search_utils.py
@@ -74,7 +74,7 @@ from mlflow.utils.search_utils import SearchUtils
                                  'value': 'RUNNING'}]),
 ])
 def test_filter(filter_string, parsed_filter):
-    assert SearchUtils._parse_search_filter(filter_string) == parsed_filter
+    assert SearchUtils.parse_search_filter(filter_string) == parsed_filter
 
 
 @pytest.mark.parametrize("filter_string, parsed_filter", [
@@ -85,7 +85,7 @@ def test_filter(filter_string, parsed_filter):
                                'key': 'm', 'value': "L'Hosp"}]),
 ])
 def test_correct_quote_trimming(filter_string, parsed_filter):
-    assert SearchUtils._parse_search_filter(filter_string) == parsed_filter
+    assert SearchUtils.parse_search_filter(filter_string) == parsed_filter
 
 
 @pytest.mark.parametrize("filter_string, error_message", [
@@ -117,7 +117,7 @@ def test_correct_quote_trimming(filter_string, parsed_filter):
 ])
 def test_error_filter(filter_string, error_message):
     with pytest.raises(MlflowException) as e:
-        SearchUtils._parse_search_filter(filter_string)
+        SearchUtils.parse_search_filter(filter_string)
     assert error_message in e.value.message
 
 
@@ -132,7 +132,7 @@ def test_error_filter(filter_string, error_message):
 ])
 def test_error_comparison_clauses(filter_string, error_message):
     with pytest.raises(MlflowException) as e:
-        SearchUtils._parse_search_filter(filter_string)
+        SearchUtils.parse_search_filter(filter_string)
     assert error_message in e.value.message
 
 
@@ -150,7 +150,7 @@ def test_error_comparison_clauses(filter_string, error_message):
 ])
 def test_bad_quotes(filter_string, error_message):
     with pytest.raises(MlflowException) as e:
-        SearchUtils._parse_search_filter(filter_string)
+        SearchUtils.parse_search_filter(filter_string)
     assert error_message in e.value.message
 
 
@@ -165,7 +165,7 @@ def test_bad_quotes(filter_string, error_message):
 ])
 def test_invalid_clauses(filter_string, error_message):
     with pytest.raises(MlflowException) as e:
-        SearchUtils._parse_search_filter(filter_string)
+        SearchUtils.parse_search_filter(filter_string)
     assert error_message in e.value.message